Move network traffic monitoring utility to syft.generic.metrics.py
